# swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MovieFlix API",
        "description": "API responsable for managing movies from MovieFlix",
        "version": "1.0.0"
    },
    "paths": {
        "/movies": {
            "get": {
                "tags": ["Movie"],
                "summary": "Get all movies",
                "responses": {
                    "200": {
                        "description": "Returns a movie list",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Movie"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/movies/{id}": {
            "put": {
                "tags": ["Movie"],
                "summary": "Updating movie",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "description": "Movie ID to be updated",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/MovieInput"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Movie updated successfully"
                    },
                    "404": {
                        "description": "Movie not found"
                    },
                    "500": {
                        "description": "Fail to update record"
                    }
                }
            },
            "delete": {
                "tags": ["Movie"],
                "summary": "Deleting movie",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "description": "Movie ID to be deleted",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Movie deleted successfully"
                    },
                    "404": {
                        "description": "Movie not found"
                    },
                    "500": {
                        "description": "Fail to remove record"
                    }
                }
            }
        },
        "/movies/{genre}": {
            "get": {
                "tags": ["Movie"],
                "summary": "Filter movies by genre",
                "parameters": [
                    {
                        "name": "genre",
                        "in": "path",
                        "required": true,
                        "description": "Movie genre name",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Movie"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Fail to find genre"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Movie": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "description": "Movie ID"
                    },
                    "title": {
                        "type": "string",
                        "description": "Movie title"
                    },
                    "genre_id": {
                        "type": "integer",
                        "description": "Movie genre ID"
                    },
                    "genres": {
                        "type": "object",
                        "properties": {
                            "id": {
                                "type": "integer",
                                "description": "Genre ID"
                            },
                            "name": {
                                "type": "string",
                                "description": "Movie genre"
                            }
                        }
                    },
                    "language_id": {
                        "type": "integer",
                        "description": "Movie language ID"
                    },
                    "languages": {
                        "type": "object",
                        "properties": {
                            "id": {
                                "type": "integer",
                                "description": "Language ID"
                            },
                            "name": {
                                "type": "string",
                                "description": "Movie Language"
                            }
                        }
                    },
                    "oscar_count": {
                        "type": "integer",
                        "description": "Movie number of Oscars"
                    },
                    "release_date": {
                        "type": "string",
                        "format": "date",
                        "description": "Movie date release"
                    }
                }
            },
            "MovieInput": {
                "type": "object",
                "required": [
                    "title",
                    "genre_id",
                    "language_id",
                    "oscar_count",
                    "release_date"
                ],
                "properties": {
                    "title": {
                        "type": "string",
                        "description": "Movie title"
                    },
                    "genre_id": {
                        "type": "integer",
                        "description": "Movie genre ID"
                    },
                    "language_id": {
                        "type": "integer",
                        "description": "Movie language ID"
                    },
                    "oscar_count": {
                        "type": "integer",
                        "description": "Movie number of Oscars"
                    },
                    "release_date": {
                        "type": "string",
                        "format": "date",
                        "description": "Movie date release"
                    }
                }
            }
        }
    }
}